updated the test for sigma space
